Flask_pH_Predict/requirements.txt
flask
flask-cors
requests
supabase
postgrest
gotrue
numpy
pandas
scikit-learn==1.6.1
tensorflow
joblib
keras
paho-mqtt